core/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.2.1
djangorestframework==3.14.0
drf-spectacular==0.26.2
et-xmlfile==1.1.0
geographiclib==2.0
geopy==2.3.0
gunicorn==20.1.0
inflection==0.5.1
jsonschema==4.17.3
kombu==5.2.4
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.1
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
redis==4.5.5
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
vine==5.0.0
wcwidth==0.2.6
